requirements.txt
requests==2.28.0
google-cloud-storage==1.26.0
pyyaml #doesn't need version
pandas
google-cloud-pubsub==2.7.0
fsspec
gcsfs


#Need to write a script as some point to scan everyone's "main.py" code and create this requirements.txt file so it doesn't need to be done every single piece of code.
